Initial commit. It's a bit ways in, but it's still in early stages.
Long explanation: The game currently lets you click a button to increment a money counter. When you have $20, you can click another button to buy a cursor, which increments another element. This element will give $1/s/u. The second buy button greys out if you can't buy it.
The UI scales depending on the screen size, but it'll be slightly off-screen if you're going any wider than 16:9 (that's just because it's justified poorly).

Things I need to do:
Classes are messy, and I need to clean and unify their systems up right now.
Add a progress bar to the scene so that you know when it'll update.
Make it so that more cursors cost more money each time.
Create a UI element for that.